Edit the Java println yasnippet.
The behavior of the println snippet with the String and the newline was
bothering me. I would delete it most of the time when I activated the
snippet, 1) because I wasn't going to print a string, and 2) I didn't
want to insert a newline.

This new snippet is much simpler and fits the way I use it.