Remove magit-emacsclient-executable configuration.
Looks like this variable name isn't used anymore. As of the commit
linked below, the name has changed but the feature still exists.

In any case, it looks like the latest version of magit solved the issue
I had that was linked in the config about commit messages opening up in
a new frame, so I can remove this configuration altogether.

Yay!

Direct GitHub link:
https://github.com/magit/magit/commit/ccb75665de3ac0913589a49e14cee077507906c4