Remove use of 's' string library.
Config did not install these beforehand, so an error happens on a
clean install.

The functions I wanted to use were simple enough so I decided to use
built-in function instead of using s. Looks like string-trim is part
of subr-x.el, which is a lesser version of subr.el. I am not sure if
the functions in that file are going away any time soon. It's an
official part of GNU Emacs so my guess is it's here to stay for now.

Require subr-x or there's a warning on start up about an undefined
string-trim function.